Move ZSN to multivariate.py
